move WarcIterator towarc_reader.py
